requirements.txt
docling
langchain-huggingface
accelerate
tiktoken==0.9.0
langchain-openai
langchain-community 
faiss-cpu
groq
toml
langchain-groq
gradio==5.26.0
markdown
unstructured[all-docs]
pillow 
lxml 
pillow




